+++ backend/fisio_find/files/models.py
@@ -1,7 +1,6 @@
 from django.db import models
 from django.conf import settings
 import boto3
-# from django.core.exceptions import ValidationError
 from treatments.models import Treatment
 
 
@@ -38,3 +37,35 @@ class PatientFile(models.Model):
     def file_url(self):
         """Devuelve la URL del archivo en DigitalOcean Spaces"""
         return f"{settings.DIGITALOCEAN_ENDPOINT_URL}/{settings.DIGITALOCEAN_SPACE_NAME}/{self.file_key}"
+
+
+class Video(models.Model):
+    treatment = models.ForeignKey(Treatment, on_delete=models.CASCADE, related_name='treatment_videos')
+    title = models.CharField(max_length=255)
+    description = models.TextField(blank=True, null=True)
+    uploaded_at = models.DateTimeField(auto_now_add=True)
+    file_key = models.CharField(
+        max_length=500, unique=True
+    )
+
+    def __str__(self): 
+        return self.title
+
+    def delete_from_storage(self):
+        """Elimina el archivo de DigitalOcean Spaces"""
+        s3_client = boto3.client(
+            "s3",
+            aws_access_key_id=settings.DIGITALOCEAN_ACCESS_KEY_ID,
+            aws_secret_access_key=settings.DIGITALOCEAN_SECRET_ACCESS_KEY,
+            endpoint_url=settings.DIGITALOCEAN_ENDPOINT_URL
+        )
+
+        try:
+            s3_client.delete_object(Bucket=settings.DIGITALOCEAN_SPACE_NAME, Key=self.file_key)
+        except Exception as e:
+            print(f"Error al eliminar el archivo de Spaces: {e}")
+
+    @property
+    def file_url(self):
+        """Devuelve la URL pública del archivo almacenado en DigitalOcean"""
+        return f"https://fisiofind-repo.fra1.digitaloceanspaces.com/{self.file_key}"
