+++ backend/fisio_find/users/serializers.py
@@ -12,9 +12,8 @@ from django.conf import settings
 from django.core.exceptions import ValidationError
 from django.core.validators import RegexValidator
 from datetime import date, datetime
-import boto3, uuid
 from users.util import validate_dni_match_letter, codigo_postal_no_mide_5, telefono_no_mide_9, validate_dni_structure
-from .models import AppUser, Patient, Physiotherapist, PhysiotherapistSpecialization, Specialization, Video, PatientFile
+from .models import AppUser, Patient, Physiotherapist, PhysiotherapistSpecialization, Specialization
 
 
 class AppUserSerializer(serializers.ModelSerializer):
@@ -518,119 +517,3 @@ class PatientAdminViewSerializer(serializers.ModelSerializer):
     class Meta:
         model = Patient
         fields = '__all__'
-
-
-class VideoSerializer(serializers.ModelSerializer):
-    file = serializers.FileField(write_only=True)  # Para recibir el archivo en el request
-    file_url = serializers.SerializerMethodField()  # Para devolver la URL pública
-    patients = serializers.PrimaryKeyRelatedField(many=True, queryset=Patient.objects.all())  
-
-    class Meta:
-        model = Video
-        fields = ["id", "patients", "title", "description", "file", "file_key", "file_url", "uploaded_at"]
-        extra_kwargs = {
-            "file_key": {"read_only": True},  # El usuario no debe enviar esto
-        }
-
-    def validate_file(self, file):
-        """Valida que el archivo sea un video permitido."""
-        allowed_extensions = (".mp4")
-        if not file.name.lower().endswith(allowed_extensions):
-            raise serializers.ValidationError("Solo se permiten archivos .mp4 .")
-        return file
-
-    def create(self, validated_data):
-        """Sube el archivo a DigitalOcean Spaces y guarda el Video en la BD."""
-        file = validated_data.pop("file")  # Extraer archivo del request
-        physiotherapist = self.context["request"].user.physio  # Usuario logueado
-        patients = validated_data.pop("patients", [])
-
-        # Generar un nombre único para evitar sobrescribir archivos
-        file_extension = file.name.split(".")[-1]
-        file_key = f"videos/{physiotherapist.id}/{uuid.uuid4()}.{file_extension}"
-
-        # Conectar a DigitalOcean Spaces
-        s3_client = boto3.client(
-            "s3",
-            region_name=settings.DIGITALOCEAN_REGION,
-            endpoint_url=settings.DIGITALOCEAN_ENDPOINT_URL,
-            aws_access_key_id=settings.DIGITALOCEAN_ACCESS_KEY_ID,
-            aws_secret_access_key=settings.DIGITALOCEAN_SECRET_ACCESS_KEY,
-        )
-
-        # Intentar subir el archivo
-        try:
-            s3_client.upload_fileobj(
-                file,
-                settings.DIGITALOCEAN_SPACE_NAME,
-                file_key,
-                ExtraArgs={"ACL": "public-read"},  # Permitir acceso público
-            )
-        except Exception as e:
-            raise serializers.ValidationError(f"Error al subir archivo: {str(e)}")
-
-        # Guardar en la BD
-        video = Video.objects.create(
-            file_key=file_key,
-            physiotherapist=physiotherapist,
-            **validated_data  # Agrega título, descripción, etc.
-        )
-        video.patients.set(patients)
-        return video
-
-    def update(self, instance, validated_data):
-        """Actualiza un video en DigitalOcean Spaces y la BD."""
-        file = validated_data.pop("file", None)  # Nuevo archivo (opcional)
-        patients = validated_data.pop("patients", None)  # Lista de pacientes (opcional)
-
-        # Actualizar título y descripción si están en los datos
-        instance.title = validated_data.get("title", instance.title)
-        instance.description = validated_data.get("description", instance.description)
-
-        if file:
-            # Conectar a DigitalOcean Spaces
-            s3_client = boto3.client(
-                "s3",
-                region_name=settings.DIGITALOCEAN_REGION,
-                endpoint_url=settings.DIGITALOCEAN_ENDPOINT_URL,
-                aws_access_key_id=settings.DIGITALOCEAN_ACCESS_KEY_ID,
-                aws_secret_access_key=settings.DIGITALOCEAN_SECRET_ACCESS_KEY,
-            )
-
-            # Borrar archivo anterior de DigitalOcean
-            try:
-                s3_client.delete_object(
-                    Bucket=settings.DIGITALOCEAN_SPACE_NAME,
-                    Key=instance.file_key,
-                )
-            except Exception as e:
-                raise serializers.ValidationError(f"Error al eliminar archivo anterior: {str(e)}")
-
-            # Generar nuevo nombre único
-            file_extension = file.name.split(".")[-1]
-            new_file_key = f"videos/{instance.physiotherapist.id}/{uuid.uuid4()}.{file_extension}"
-
-            # Subir nuevo archivo
-            try:
-                s3_client.upload_fileobj(
-                    file,
-                    settings.DIGITALOCEAN_SPACE_NAME,
-                    new_file_key,
-                    ExtraArgs={"ACL": "public-read"},
-                )
-            except Exception as e:
-                raise serializers.ValidationError(f"Error al subir nuevo archivo: {str(e)}")
-
-            # Actualizar referencia en la BD
-            instance.file_key = new_file_key
-
-        # Si se pasan pacientes, actualizarlos
-        if patients is not None:
-            instance.patients.set(patients)
-
-        instance.save()
-        return instance
-
-    def get_file_url(self, obj):
-        """Devuelve la URL completa del archivo."""
-        return f"{settings.DIGITALOCEAN_ENDPOINT_URL}/{obj.file_key}"
