+++ backend/fisio_find/users/urls.py
@@ -20,12 +20,6 @@ urlpatterns = [
     path('profile/', PatientProfileView.as_view(), name='profile'),
     path('current-user/', return_user, name='current_user'),
 
-    path('videos/upload/', create_video, name='upload_video'),
-    path('videos/delete/<int:video_id>/', delete_video, name='delete_video'),
-    path('videos/list-my-videos/', list_my_videos, name='list_my_videos'),
-    path('videos/stream-video/<int:video_id>/', stream_video, name='stream_video'),
-    path('videos/update-video/<int:video_id>/', update_video, name='update_video'),
-
     path('services/<int:physio_id>/', physio_get_services_view, name='physio_get_xservices'),
