+++ backend/fisio_find/users/views.py
@@ -3,10 +3,7 @@ import stripe
 import os
 from django.conf import settings
 import json
-import boto3
 from django.shortcuts import get_object_or_404
-from django.conf import settings
-from django.http import HttpResponse, StreamingHttpResponse
 from django.db.models import Q
 from rest_framework.decorators import api_view, permission_classes
 from rest_framework.response import Response
@@ -29,10 +26,9 @@ from .serializers import (
     PhysioSerializer,
     PatientSerializer,
     AppUserSerializer,
-    VideoSerializer,
     PhysioUpdateSerializer
     )
-from .models import Physiotherapist, Patient, AppUser, Video
+from .models import Physiotherapist, Patient, AppUser
 from .permissions import (
     IsPatient,
     IsPhysiotherapist,
@@ -575,227 +571,3 @@ class AdminPatientDelete(generics.DestroyAPIView):
     queryset = Patient.objects.all()
     serializer_class = PatientRegisterSerializer
 """
-
-@api_view(['POST'])
-@permission_classes([IsPhysiotherapist])
-def create_video(request):
-    print("🔍 Datos recibidos:", request.data)
-
-    # Convertir request.data en un diccionario mutable (para modificar el QueryDict)
-    mutable_data = request.data.copy()
-
-    # Manejo de `patients` usando emails en lugar de IDs
-    patients_raw = mutable_data.get("patients")
-    if patients_raw:
-        if isinstance(patients_raw, str):
-            try:
-                # Se espera una cadena JSON con una lista de emails, por ejemplo: '["email1@example.com", "email2@example.com"]'
-                patients_list = json.loads(patients_raw)
-                if isinstance(patients_list, list) and all(isinstance(i, str) for i in patients_list):
-                    # Buscar usuarios que coincidan con los emails proporcionados
-                    users = AppUser.objects.filter(email__in=patients_list)
-                    found_emails = set(users.values_list("email", flat=True))
-                    missing_emails = set(patients_list) - found_emails
-                    if missing_emails:
-                        return Response(
-                            {"errorManaged": f"Usuarios no encontrados para emails: {', '.join(list(missing_emails))}"},
-                            status=status.HTTP_400_BAD_REQUEST
-                        )
-                    
-                    # Obtener los pacientes asociados a esos usuarios
-                    patients = Patient.objects.filter(user__in=users)
-                    patients_found_emails = set(patients.values_list("user__email", flat=True))
-                    missing_patients = found_emails - patients_found_emails
-                    if missing_patients:
-                        return Response(
-                            {"errorManaged": f"Paciente no encontrado para usuarios con emails: {', '.join(list(missing_patients))}"},
-                            status=status.HTTP_400_BAD_REQUEST
-                        )
-                    
-                    # Convertir el queryset a una lista de IDs de Patient
-                    patient_ids = list(patients.values_list("id", flat=True))
-                    mutable_data.setlist("patients", patient_ids)
-                else:
-                    return Response(
-                        {"errorManaged": "Formato de patients incorrecto, debe ser una lista de emails"},
-                        status=status.HTTP_400_BAD_REQUEST
-                    )
-            except json.JSONDecodeError:
-                return Response(
-                    {"errorManaged": "Formato de patients inválido"},
-                    status=status.HTTP_400_BAD_REQUEST
-                )
-
-    print("📌 Datos después de procesar:", mutable_data)  # Para depuración
-
-    # Pasamos mutable_data en lugar de request.data al serializer
-    serializer = VideoSerializer(data=mutable_data, context={"request": request})
-
-    if serializer.is_valid():
-        video = serializer.save()
-        return Response(
-            {
-                "message": "Archivo creado correctamente",
-                "video": VideoSerializer(video).data
-            },
-            status=status.HTTP_201_CREATED
-        )
-
-    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-
-@api_view(['DELETE'])
-@permission_classes([IsPhysiotherapist])
-def delete_video(request, video_id):
-    user = request.user
-    print(user.physio.id)  # Depuración
-    print(f"Usuario autenticado: {user}, Rol: {getattr(user, 'physiotherapist', None)}")  # Depuración
-    try:
-        video = Video.objects.get(id=video_id)
-        print(f"Video encontrado: {video}")  # Depuración
-    
-        if not hasattr(user, 'physio') or video.physiotherapist.id != user.physio.id:
-            return Response({"error": "No tienes permiso para eliminar este video"}, status=status.HTTP_403_FORBIDDEN)
-
-        video.delete_from_storage()
-        video.delete()
-
-        return Response({"message": "Video eliminado correctamente"}, status=status.HTTP_200_OK)
-
-    except Video.DoesNotExist:
-        return Response({"error": "Video no encontrado"}, status=status.HTTP_404_NOT_FOUND)
-    
-
-@api_view(['GET'])
-@permission_classes([IsPhysioOrPatient])  
-def list_my_videos(request):
-    user = request.user
-
-    try:
-        if hasattr(user, 'patient'):
-            print("Patient:", user.patient.id)
-            videos = Video.objects.filter(patients__id=user.patient.id)
-
-        elif hasattr(user, 'physio'):
-            print("Physio:", user.physio.id)
-            videos = Video.objects.filter(physiotherapist=user.physio.id)
-
-        else:
-            return Response({"error": "No tienes permisos para ver estos videos"}, status=status.HTTP_403_FORBIDDEN)
-
-        serializer = VideoSerializer(videos, many=True)
-        return Response(serializer.data, status=status.HTTP_200_OK)
-
-    except Exception as e:
-        print(f"Error al obtener los videos: {e}")
-        return Response({"error": "Hubo un problema al obtener los videos"}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-
-
-s3_client = boto3.client(
-    's3',
-    aws_access_key_id=settings.DIGITALOCEAN_ACCESS_KEY_ID,
-    aws_secret_access_key=settings.DIGITALOCEAN_SECRET_ACCESS_KEY,
-    region_name=settings.DIGITALOCEAN_REGION,
-    endpoint_url=settings.DIGITALOCEAN_ENDPOINT_URL,
-)
-
-
-@api_view(['GET'])
-@permission_classes([IsPatient])
-def stream_video(request, video_id):
-    try:
-        video = Video.objects.get(id=video_id)
-        if not hasattr(request.user, "patient"):
-            return Response({'error': 'No tienes un perfil de paciente'}, status=403)
-
-        patient_id = request.user.patient.id  # Obtener el ID del paciente
-
-        if patient_id not in video.patients.values_list('id', flat=True):
-            return Response({'error': 'No tienes acceso a este video'}, status=403)
-
-        video_object = s3_client.get_object(
-            Bucket=settings.DIGITALOCEAN_SPACE_NAME,
-            Key=video.file_key
-        )
-
-        video_size = video_object["ContentLength"]
-        video_body = video_object["Body"]
-
-        # Manejar streaming por fragmentos (range requests)
-        range_header = request.headers.get("Range", None)
-        if range_header:
-            range_value = range_header.replace("bytes=", "").split("-")
-            start = int(range_value[0]) if range_value[0] else 0
-            end = int(range_value[1]) if len(range_value) > 1 and range_value[1] else video_size - 1
-            chunk_size = end - start + 1
-
-            # Leer solo el fragmento necesario
-            video_body.seek(start)
-            video_chunk = video_body.read(chunk_size)
-
-            # Responder con `206 Partial Content`
-            response = HttpResponse(video_chunk, content_type="video/mp4")
-            response["Content-Range"] = f"bytes {start}-{end}/{video_size}"
-            response["Accept-Ranges"] = "bytes"
-            response["Content-Length"] = str(chunk_size)
-            response["Cache-Control"] = "no-cache"
-            response["Connection"] = "keep-alive"
-            response.status_code = 206
-        else:
-            # Streaming completo en fragmentos
-            def stream_file():
-                for chunk in video_body.iter_chunks():
-                    yield chunk
-
-            response = StreamingHttpResponse(stream_file(), content_type="video/mp4")
-            response["Content-Length"] = str(video_size)
-            response["Accept-Ranges"] = "bytes"
-            response["Cache-Control"] = "no-cache"
-            response["Connection"] = "keep-alive"
-        
-        return response
-
-    except Video.DoesNotExist:
-        return Response({"error": "El video no existe"}, status=404)
-    except Exception as e:
-        return Response({"error": str(e)}, status=500)
-
-
-@api_view(['PUT'])
-@permission_classes([IsPhysiotherapist])  # Solo usuarios autenticados pueden acceder
-def update_video(request, video_id):
-    try:
-        # Obtener el video desde la BD
-        video = Video.objects.get(id=video_id)
-
-        # Verificar que el usuario autenticado es el fisioterapeuta propietario del video
-        if request.user.physio != video.physiotherapist:
-            return Response({'error': 'No tienes permiso para actualizar este video'}, status=status.HTTP_403_FORBIDDEN)
-
-    except Video.DoesNotExist:
-        return Response({'error': 'El video no existe'}, status=status.HTTP_404_NOT_FOUND)
-
-    # Convertir request.data en un diccionario mutable
-    mutable_data = request.data.copy()
-
-    # Manejo de `patients`
-    patients_raw = mutable_data.get("patients")
-    if patients_raw:
-        if isinstance(patients_raw, str):
-            try:
-                patients_list = json.loads(patients_raw)  # Convierte "[1, 3]" a [1, 3]
-            except json.JSONDecodeError:
-                return Response({"error": "Formato de patients inválido"}, status=status.HTTP_400_BAD_REQUEST)
-        elif isinstance(patients_raw, list):
-            patients_list = patients_raw
-        else:
-            return Response({"error": "Formato de patients incorrecto, debe ser una lista de enteros"}, status=status.HTTP_400_BAD_REQUEST)
-        
-    # Serializar con los datos nuevos
-    serializer = VideoSerializer(video, data=mutable_data, partial=True, context={'request': request})
-
-    if serializer.is_valid():
-        serializer.save()
-        return Response({"message": "Video actualizado correctamente", "video": serializer.data}, status=status.HTTP_200_OK)
-    
-    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
